store string length as size_t - not unsigned

+++ src/main.cc
@@ -494,7 +494,7 @@ void CallTestFunction(bool immediately = false) {
 		auto b = HandleTopLevelExpression(std::move(call_expr), true);
 		char* buf;
 		char* volvoxstrTestFunction;
-		unsigned lalloc;
+		size_t lalloc;
 		cstr2volvoxstr(volvoxstrTestFunction, lalloc, buf, TestFunction);
 		showtestres(1, 79, volvoxstrTestFunction, b);
 		return;
